+++ DragonShield/helpers/Color+Palette.swift
@@ -1,0 +1,7 @@
+import SwiftUI
+
+extension Color {
+    static let success = Color(red: 30/255, green: 142/255, blue: 62/255)
+    static let warning = Color(red: 224/255, green: 168/255, blue: 0/255)
+    static let error = Color(red: 211/255, green: 47/255, blue: 47/255)
+}